// src/features/users/domain/errors/user-error-code.enum.ts
export enum UserErrorCode {
  USER_NOT_FOUND = 'USER_NOT_FOUND',
  EMAIL_ALREADY_EXISTS = 'EMAIL_ALREADY_EXISTS',
  USERNAME_ALREADY_EXISTS = 'USERNAME_ALREADY_EXISTS',
  INVALID_CURSOR = 'INVALID_CURSOR',
  INVALID_VERIFICATION_CODE = 'INVALID_VERIFICATION_CODE',
  EXPIRED_VERIFICATION_CODE = 'EXPIRED_VERIFICATION_CODE',
  ALREADY_VERIFIED = 'ALREADY_VERIFIED',
  USER_ALREADY_SUSPENDED = 'USER_ALREADY_SUSPENDED',
  INVALID_STATUS_TRANSITION = 'INVALID_STATUS_TRANSITION'
}